Added other yelp services to iPhone

+++ Fall2014IOSApp/MenuViewController.m
@@ -73,7 +73,7 @@
 #pragma mark - RNFrostedSidebarDelegate
 
 - (void)sidebar:(RNFrostedSidebar *)sidebar didTapItemAtIndex:(NSUInteger)index {
-    NSLog(@"Tapped item at index %lu",index);
+    NSLog(@"Tapped item at index %lu",(unsigned long)index);
     if (index == 0) {
         
         [self performSegueWithIdentifier:@"segueToConfSched" sender:self];
