+++ Fall2014IOSApp/StartPageViewController.m
@@ -452,7 +452,7 @@
 #pragma mark - RNFrostedSidebarDelegate
 
 - (void)sidebar:(RNFrostedSidebar *)sidebar didTapItemAtIndex:(NSUInteger)index {
-    NSLog(@"Tapped item at index %lu",index);
+    NSLog(@"Tapped item at index %lu",(unsigned long)index);
     switch (index) {
         case 0:
             [self performSegueWithIdentifier:@"segueToHotel" sender:self];
